ascii: AsciiCanvas
Port components/ascii/ascii-canvas.tsx → Svelte 5.

- Extract shared pixel utilities (sampleImage, renderAsciiFrame, getGrid,
  lumToTone) into src/lib/ascii/ascii-utils.ts
- Canvas resize + image-load logic moved to Svelte action
  src/lib/actions/ascii-canvas.ts (legacy update/destroy API, plain .ts)
- AsciiCanvas.svelte: thin wrapper div + canvas with use:asciiCanvas
- Autofixer: no issues
